refactor: move _implement_latex_subscript to ampform.sympy
